Add info to Membarr page
This commit adds instructions for setting up Membarr - a bot that invites Discord users to Plex and Jellyfin and automates invites with roles. The commit includes details on Membarr's features, a project home, and setup instructions for creating a Discord bot, installing Membarr, setting up Plex parameters, and getting a Plex token.