Add Membarr Docs (#205)
* Add new app Membarr to sandbox

This commit adds Membarr, a new app to the Sandbox section of the documentation. It is now included in the navigation and can be accessed from the sidebar.

* Add info to Membarr page

This commit adds instructions for setting up Membarr - a bot that invites Discord users to Plex and Jellyfin and automates invites with roles. The commit includes details on Membarr's features, a project home, and setup instructions for creating a Discord bot, installing Membarr, setting up Plex parameters, and getting a Plex token.