Added Homeassistant documentation page
Introduced a new documentation page for Homeassistant within the sandbox apps section of the docs. The page outlines the purpose of Homeassistant, provides important links such as project home, docs, GitHub, and Docker, and details steps for installation and accessing the web interface. It advises on the complexity of remote server setup and informs that the role isn't protected by default Authelia/SSO middleware.