Updated Home Assistant doc with Nabu Casa info
Added a note to the Home Assistant documentation clarifying that while Nabu Casa is not necessary for remote access, it is an available paid service that supports the project. Further emphasized that the current setup is prepared for use with a reverse proxy as an alternative to Nabu Casa for remote access.