Add Home Assistant MQTT and Node-Red documentation (#272)
* Added Homeassistant documentation page

Introduced a new documentation page for Homeassistant within the sandbox apps section of the docs. The page outlines the purpose of Homeassistant, provides important links such as project home, docs, GitHub, and Docker, and details steps for installation and accessing the web interface. It advises on the complexity of remote server setup and informs that the role isn't protected by default Authelia/SSO middleware.

* Update Home Assistant documentation link

* Update Home Assistant setup documentation

* Update Home Assistant setup instructions

* Add Homeassistant to sandbox apps

* Added Home Assistant documentation under Home Automation

Extended the documentation navigation to include a new section on Home Automation--specifically, the addition of Home Assistant. This update makes it easier for users to find information on integrating Home Automation solutions with our platform.

* Expand Home Automation section with MQTT and Node-Red docs

Added documentation links for MQTT and Node-Red to the Home Automation subsection in the navigation of the project's MkDocs configuration. This update ensures users have access to a wider range of resources for setting up and managing their home automation systems.

* Normalize 'Homeassistant' casing, add MQTT and Node-RED

Standardized the capitalization of 'Homeassistant' to 'homeassistant' for consistency across our documentation. Also introduced documentation links for MQTT and Node-RED to the sandbox apps index, expanding our coverage of supported applications.

* yup

* Updated MQTT and Node-RED documentation

Enhanced the MQTT installation instructions with corrected code block formatting for better readability. Added setup instructions to the Node-RED documentation to include a palette for integration with Home Assistant, facilitating easier connection setups for users.

* Add MQTT project details and documentation links

* Update Home Assistant and MQTT docs, clarify Node Red

- Clarified that Home Assistant has its own authentication system and added a new section for Home Assistant Community Store (HACS) addons.
- Removed redundant setup instructions in MQTT documentation and added a note regarding the lack of a web interface.
- Corrected Node Red description to reflect its actual use as a visual programming tool. Adjusted warning to reflect that Node Red is now protected by Authelia/SSO middleware and reorganized setup instructions for clearer readability.

* Updated Home Assistant doc with Nabu Casa info

Added a note to the Home Assistant documentation clarifying that while Nabu Casa is not necessary for remote access, it is an available paid service that supports the project. Further emphasized that the current setup is prepared for use with a reverse proxy as an alternative to Nabu Casa for remote access.

* Update MQTT and Node-Red documentation links

Updated the documentation to reflect the accurate URLs and project information for MQTT (Eclipse Mosquitto) and Node-Red. Enhanced the MQTT section with clearer setup instructions, including security best practices, and corrected links to MQTT's official resources. For Node-Red, clarified the installation and setup guide, and provided updated URLs to Node-Red's home, documentation, and Docker hub. These changes ensure users have the latest information and a more secure, streamlined setup process for both applications.